.automatically/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
oauthlib==3.2.0
python-frontmatter==1.0.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
tweepy==4.10.0
urllib3==1.26.9
